Update Makefile
I have made a modification to the Makefile to handle deprecated function warnings in OpenSSL 3.0 and later versions. The modification includes adding the -Wno-deprecated-declarations option to the compilation flags (CFLAGS).

By adding this option, it disables the warnings related to deprecated functions, allowing for error-free compilation within the context of OpenSSL 3.0 and newer versions.